Handle non default export in babel-plugin-zent (#1086)
* Extract constants from components

* fix babel-plugin-zent for non default exports
